explantion to each function
im going to explain each function:
play song- the function receives a song ID and needs to play the song. I created a cur variable which is going to be the text that displays each time a song is being played. also I created 2 variables that will be the song title and artist name. I take them and put them inside the inner HTML of cur and then we will know which song is being played by a specific artist.
generate songs and playlists- these 2 functions is always updating the browser display in any kind of change that happened.
remove song- the function receives a song object, find it's index in song array and using splice to delete it. after that we will use generate song to update the songs array.
create song element- receives all song keys and create a new song element using the create element function. I have add 2 button elements that would allow us to play/remove a song when we want, these button also has classes.
handle song click event- this function receives an event and I'm creating 2 variables around him. the first is IDselected which would give us the id of the song that the event occurd on him. the second is action that gives us the class of the button that we clicked on. then the function checks if action is remove or play and running the function playsong and removesong accordingly. we use IDselected to see on which song to make the action.
addsong- receives a song object and push him in to the song array and use generate song to update the array.
 handle addsong event- the function receives an event that happened by clicking the add button. the function then creates a variable for each of the song keys that were put in by the user and uses addsong function to push the song in to the array.
playlistDuration- the function receives an ID and gives back the entire playlist duration.
convert time- take a number that represents duration and return the duration in mm:ss format.
convert seconds- does the opposite from convert time.
create playlist element- receives all playlist keys. and using the id and name to create elements for them. for songs it uses playlistDuration to give us the playlist duration and uses convert time to change it to the right format. after these 2 function did their thing we create another element from what they created to represent the songs element.